[chore] 통합테스트, 단위테스트 폴더 분리
